+++ src/ComboboxWithSearch/ComboboxWithSearch.js
@@ -1,10 +1,12 @@
 import React, { Component } from 'react';
+import ReactDOM from 'react-dom';
 import PropTypes from 'prop-types';
 import { Icon } from '@opuscapita/react-icons';
 import { Async as Select } from '@opuscapita/react-select';
 import { createFilter } from 'react-select';
 
-import SearchModal from './SearchModal';
+import FloatingMenu from './FloatingMenu';
+import SearchModal from '../SearchModal';
 
 import 'react-table/react-table.css'
 import './ComboboxWithSearch.scss';
@@ -15,7 +17,7 @@ const ICON_SIZE = {
 };
 
 const valueBecomesDefinedOrIsCleared = (nextProps, currentProps) => !!nextProps.value !== !!currentProps.value;
-const OptionValueChanges = (nextProps, currentProps) => nextProps.value && currentProps.value &&
+const optionValueChanges = (nextProps, currentProps) => nextProps.value && currentProps.value &&
   nextProps.value.value !== currentProps.value.value;
 
 export class ComboboxWithSearch extends Component {
@@ -30,8 +32,12 @@ export class ComboboxWithSearch extends Component {
     };
   }
 
+  componentDidMount() {
+    this.dropdownFieldNode = ReactDOM.findDOMNode(this);
+  }
+
   componentWillReceiveProps(nextProps) {
-    if (valueBecomesDefinedOrIsCleared(nextProps, this.props) || OptionValueChanges(nextProps, this.props)) {
+    if (valueBecomesDefinedOrIsCleared(nextProps, this.props) || optionValueChanges(nextProps, this.props)) {
       this.setState({
         value: nextProps.value,
       });
@@ -82,6 +88,13 @@ export class ComboboxWithSearch extends Component {
         </div>
       );
     };
+    const Menu = props => {
+      const newProps = {
+        ...props,
+        dropdownFieldNode: this.dropdownFieldNode,
+      };
+      return FloatingMenu(newProps);
+    };
     return (
       <div className="combobox-with-search">
         <div className="combobox-with-search__combobox">
@@ -91,7 +104,10 @@ export class ComboboxWithSearch extends Component {
             loadOptions={loadOptions}
             onChange={value => this.handleChange(value)}
             value={this.state.value}
-            components={{ DropdownIndicator }}
+            components={{
+              DropdownIndicator,
+              Menu,
+            }}
             noOptionsMessage={() => (localizationTexts.noItems || '--')}
             loadingMessage={() => (localizationTexts.loading || 'Loading...')}
             filterOption={createFilter({